requirements.txt
llama-index
llama-index-core
llama-index-embeddings-openai
llama-parse
pyarrow
pandas
llama-index-postprocessor-cohere-rerank
llama-index-vector-stores-kdbai
llama-index-embeddings-voyageai
llama-index-vector-stores-pinecone
pinecone-client
kdbai_client==0.1.2
fastapi
uvicorn